requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.4
gunicorn==19.6.0
oauthlib==2.0.1
psycopg2==2.6.2
PyJWT==1.4.2
python-social-auth==0.3.3
python3-openid==3.0.10
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
whitenoise==3.2
pinax_theme_bootstrap
dj-places
django-s3direct
